Services CMS + About styling (#23)
* CMS: move all md to content + connect services to CMS

* style: Service, Nav Links
