+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "bloomer": "^0.6.3",
     "bulma": "^0.6.0",
+    "classnames": "^2.2.5",
     "font-awesome": "^4.7.0",
     "gatsby": "^1.9.63",
     "gatsby-image": "^1.0.28",
