+++ src/templates/about-page.js
@@ -1,35 +1,32 @@
-import React from 'react'
-import Helmet from 'react-helmet'
-import { Box, Column, Columns, Container, Content, Section } from 'bloomer'
-import PageHeader from '../components/page-header'
-import Img from 'gatsby-image'
+import React from 'react';
+import Helmet from 'react-helmet';
+import { Box, Column, Columns, Container, Content, Section } from 'bloomer';
+import PageHeader from '../components/page-header';
+import Img from 'gatsby-image';
 
 export default ({ data }) => {
-  const siteTitle = data.site.siteMetadata.title
-  const { markdownRemark: post } = data
+  const siteTitle = data.site.siteMetadata.title;
+  const { markdownRemark: post } = data;
   return (
     <Section>
       <Helmet title={`About | ${siteTitle}`} />
-      <Container>
-        <Columns isVCentered={true}>
-          <Column isSize={4} isOffset={1} style={{ marginRight: '2em' }}>
-            <Box>
-              <Img
-                sizes={data.file.childImageSharp.sizes}
-                alt="Caroline"
-                title="About Caroline"
-              />
-            </Box>
-          </Column>
-          <Column isSize={6} style={{ paddingRight: '1em' }}>
-            <PageHeader center title={post.frontmatter.title} />
-            <Content dangerouslySetInnerHTML={{ __html: post.html }} />
-          </Column>
-        </Columns>
+      <Container style={{ maxWidth: 800 }}>
+        <div css={{ display: 'flex', alignItems: 'center', flexDirection: 'column'}}>
+          <PageHeader center title={post.frontmatter.title} />
+
+          <Box style={{ width: 400 }}>
+            <Img
+              sizes={data.file.childImageSharp.sizes}
+              alt="Caroline"
+              title="About Caroline"
+            />
+          </Box>
+          <Content dangerouslySetInnerHTML={{ __html: post.html }} />
+        </div>
       </Container>
     </Section>
-  )
-}
+  );
+};
 
 export const query = graphql`
   query AboutPage($slug: String!) {
@@ -49,10 +46,10 @@ export const query = graphql`
     }
     file(relativePath: { eq: "cdlr0.jpg" }) {
       childImageSharp {
-        sizes(maxWidth: 450) {
+        sizes(maxWidth: 400) {
           ...GatsbyImageSharpSizes
         }
       }
     }
   }
-`
+`;
